+++ src/components/admins/DeleteUser.vue
